design solver mixin
