+++ dnachisel/DnaOptimizationProblem/mixins/__init__.py
@@ -1,9 +1,11 @@
 from .ConstraintsSolverMixin import ConstraintsSolverMixin
 from .ObjectivesMaximizerMixin import ObjectivesMaximizerMixin
 from .RecordRepresentationMixin import RecordRepresentationMixin
+from .DesignSolverMixin import DesignSolverMixin
 
 __all__ = [
     'ConstraintsSolverMixin',
     'ObjectivesMaximizerMixin',
+    'DesignSolverMixin',
     'RecordRepresentationMixin'
 ]
