+++ dnachisel/tools/DBSQLite.py
@@ -6,7 +6,7 @@ from subprocess import check_output
 from time import strftime
 from uuid import uuid4
 
-from .tools import pick_random
+from .mathtools import pick_random
 
 
 class DBSQLite():
